Use BigDecimal for currency calculations
Refactors the currency conversion logic to use `BigDecimal` instead of `Double` for improved precision in financial calculations. This change is applied across the data, domain, and presentation layers, including the `ExchangeRate` model, repository, use case, and `CurrencyExchangeViewModel`.

Also introduces a subtle UI change, updating the background color of the currency input fields to `DarkGray` for better visual separation. The displayed amounts are now formatted to two decimal places.

+++ feature/fxRatesConverter/src/main/java/com/wegielek/feature/fxRatesConverter/data/model/ExchangeRateDto.kt
@@ -1,13 +1,14 @@
 package com.wegielek.feature.fxRatesConverter.data.model
 
 import com.wegielek.feature.fxRatesConverter.domain.model.ExchangeRate
+import java.math.BigDecimal
 
 data class ExchangeRateDto(
     val from: String,
     val to: String,
-    val rate: Double,
-    val fromAmount: Double,
-    val toAmount: Double,
+    val rate: BigDecimal,
+    val fromAmount: BigDecimal,
+    val toAmount: BigDecimal,
 ) {
     fun toDomain() = ExchangeRate(from, to, rate, fromAmount, toAmount)
 }
