+++ feature/fxRatesConverter/src/main/java/com/wegielek/feature/fxRatesConverter/domain/model/ExchangeRate.kt
@@ -1,9 +1,11 @@
 package com.wegielek.feature.fxRatesConverter.domain.model
 
+import java.math.BigDecimal
+
 data class ExchangeRate(
     val from: String,
     val to: String,
-    val rate: Double,
-    val fromAmount: Double,
-    val toAmount: Double,
+    val rate: BigDecimal,
+    val fromAmount: BigDecimal,
+    val toAmount: BigDecimal,
 )
