+++ feature/fxRatesConverter/src/main/java/com/wegielek/feature/fxRatesConverter/domain/repository/ExchangeRatesRepository.kt
@@ -1,11 +1,12 @@
 package com.wegielek.feature.fxRatesConverter.domain.repository
 
 import com.wegielek.feature.fxRatesConverter.domain.model.ExchangeRate
+import java.math.BigDecimal
 
 interface ExchangeRatesRepository {
     suspend fun getRates(
         from: String,
         to: String,
-        amount: Double,
+        amount: BigDecimal,
     ): ExchangeRate
 }
